requirements.txt
requests
pandas
scikit-learn
python-dotenv
tiktoken
groq